rename `.c -> .h`
- `gsw_check_data.c -> gsw_check_data.h`

- `gsw_saar_data.c -> gsw_saar_data.h`
