+++ make_saar_data.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
 #  $id$
 """
-Make gsw_saar_data.c from the current gsw_data_v3_0.nc.  This is a developer
+Make gsw_saar_data.h from the current gsw_data_v3_0.nc.  This is a developer
 utility and not a part of the public distribution, but its end-product is.
-Note that it generates gsw_saar_data.c but will not overwrite it if it exists.
+Note that it generates gsw_saar_data.h but will not overwrite it if it exists.
 General concept: we don't want end-users of this distribution to require having
 netcdf installed, nor do we want to incur the I/O overhead every time this
 library is used.  So we simply generate static data from the netcdf file that
@@ -42,9 +42,9 @@ vars = [["p_ref", "", [nz]], ["lats_ref", "", [ny]], ["longs_ref", "", [nx]],
         ["saar_ref", "SAAR_ref", [nx,ny,nz]],
         ["delta_sa_ref", "deltaSA_ref", [nx,ny,nz]],["ndepth_ref", "", [nx,ny]]]
 
-fname = "gsw_saar_data.c"
+fname = "gsw_saar_data.h"
 if os.path.exists(fname):
-    print("Will not overwrite existing gsw_saar_data.c. Exiting.")
+    print("Will not overwrite existing gsw_saar_data.h. Exiting.")
     sys.exit(1)
 
 out = open(fname, "w")
